Subscription example to run on express Use console for now to print the subscription messages
