+++ example/example.js
@@ -16,21 +16,8 @@
  */
 
 'use strict';
-var Publisher = require('mongo-subs');
-var commandline = require('command-line-args');
 
-var cli = commandline([
-  { name: 'mongouri', alias: 'm', type: String },
-  { name: 'mongodbname', alias: 'd', type: String }
-]);
-
-var options = cli.parse();
-var mongouri = options.mongouri || '127.0.0.1:27017';
-var mongodbname = options.mongodbname || 'mongodbapi';
-
-var publisher = Publisher.createWith('mongodb://' + options.mongouri + '/local', {
-  ns: options.mongodbname, coll: 'oplog.$main'
-});
+var publisher = require('./publisher');
 
 console.log('Listening on mongo:' + mongouri + ' dbname: ' + mongodbname);
