+++ example/package.json
@@ -5,10 +5,13 @@
   "author": "Vignesh Periasami",
   "dependencies": {
     "command-line-args": "^2.1.6",
+    "debug": "^2.2.0",
     "express": "^4.13.4",
+    "jade": "^1.11.0",
+    "jquery": "^2.2.1",
     "mongo-oplog": "^0.2.1",
     "mongo-subs": "file:../lib",
-    "socket.io": "^1.3.6",
-    "jade": "^1.11.0"
+    "socket.io": "^1.4.5",
+    "socket.io-client": "^1.4.5"
   }
 }
