+++ example/views/pages/hellosubs.jade
@@ -2,6 +2,7 @@ doctype html
 html(lang='en')
   head
     title Subscription Demo
+    script(src='/static/dist/mongolisten.js', type='text/javascript')
 
   body
     h3 Hello subscription
