+++ lib/index.js
@@ -1,3 +1,4 @@
 'use strict';
 
-module.exports = require('./src/publisher');
+var exports = module.exports = require('./src/publisher');
+exports.MongoRPC = require('./src/mongosubsrpc');
