+++ lib/package.json
@@ -1,6 +1,6 @@
 {
   "name": "mongo-subs",
-  "version": "1.0.0",
+  "version": "1.0.3",
   "description": "Mongo subscription library",
   "author": "Vignesh Periasami",
   "files": [
@@ -9,7 +9,6 @@
   ],
   "browser": "./src/clientmongorpc.js",
   "dependencies": {
-    "mongo-oplog": "^0.2.1",
-    "socket.io": "^1.3.6"
+    "mongo-oplog": "^0.2.1"
   }
 }
