+++ package.json
@@ -6,6 +6,7 @@
   "devDependencies": {
     "chai": "^3.2.0",
     "gulp": "^3.9.0",
+    "gulp-browserify": "^0.5.1",
     "gulp-eslint": "^1.0.0",
     "gulp-mocha": "^2.1.3",
     "mocha": "^2.3.2",
